requirements.txt
streamlit
openai
pandas
numpy
Pillow
matplotlib
scikit-learn
